refactor: service에서 수행하던 dto 생성 로직을 dto 내부로 이관
