+++ src/main/java/service/ChessGameService.java
@@ -16,15 +16,11 @@ import dto.GameTurnDto;
 import dto.PieceDto;
 import dto.PositionDto;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class ChessGameService {
-
-    public static final String DOT = ".";
 
     private final PieceDao pieceDao;
     private final GameTurnDao gameTurnDao;
@@ -50,6 +46,29 @@ public class ChessGameService {
         }
     }
 
+    private GameTurnDto makeGameTurnDto(ChessBoard chessBoard) {
+        return new GameTurnDto(chessBoard.getTurnOfColor().name());
+    }
+
+    public ChessBoardStateDto makeChessBoardState() {
+        return new ChessBoardStateDto(findChessBoard());
+    }
+
+    public ChessGameScoreDto calculateScore() {
+        ChessBoard chessBoard = findChessBoard();
+        ScoreCalculator scoreCalculator = new ScoreCalculator(chessBoard.getBlackPieces(), chessBoard.getWhitePieces());
+        return new ChessGameScoreDto(scoreCalculator);
+    }
+
+    public boolean isGameEnded() {
+        return findChessBoard().isGameEnded();
+    }
+
+    public void deleteChessBoard() {
+        pieceDao.deleteAllPieces();
+        gameTurnDao.delete();
+    }
+
     private ChessBoard findChessBoard() {
         List<PieceDto> pieceDtos = pieceDao.findAllPieces();
 
@@ -59,7 +78,7 @@ public class ChessGameService {
         }
 
         Map<Position, Piece> loadBoard = new HashMap<>();
-        
+
         for (PieceDto pieceDto : pieceDtos) {
             PieceName pieceName = PieceName.of(pieceDto.getName());
             Color pieceColor = Color.valueOf(pieceDto.getPieceColor());
@@ -72,54 +91,12 @@ public class ChessGameService {
 
         GameTurnDto gameTurnDto = gameTurnDao.find();
         Color turnOfColor = Color.valueOf(gameTurnDto.getTurnOfColor());
-        
+
         return new ChessBoard(loadBoard, turnOfColor);
     }
 
     private Piece makePiece(PieceName pieceName, Color pieceColor) {
         PieceMaker pieceMaker = PieceMaker.from(pieceName);
         return pieceMaker.make(pieceColor);
-    }
-
-    private GameTurnDto makeGameTurnDto(ChessBoard chessBoard) {
-        return new GameTurnDto(chessBoard.getTurnOfColor().name());
-    }
-
-    public ChessGameScoreDto calculateScore() {
-        ChessBoard chessBoard = findChessBoard();
-        ScoreCalculator scoreCalculator = new ScoreCalculator(chessBoard.getBlackPieces(), chessBoard.getWhitePieces());
-        return new ChessGameScoreDto(scoreCalculator.getBlackScore(), scoreCalculator.getWhiteScore());
-    }
-
-    public ChessBoardStateDto makeChessBoardState() {
-        Map<Position, Piece> chessBoard = findChessBoard().getChessBoard();
-        List<String> boardState = new ArrayList<>();
-
-        for (List<Position> rowPositions : Position.getAllPositions()) {
-            boardState.add(makeRowState(chessBoard, rowPositions));
-        }
-        return new ChessBoardStateDto(boardState);
-    }
-
-    private String makeRowState(Map<Position, Piece> chessBoard, List<Position> rowPositions) {
-        return rowPositions.stream()
-                .map(position -> convertPieceNameOrDot(chessBoard, position))
-                .collect(Collectors.joining());
-    }
-
-    private String convertPieceNameOrDot(Map<Position, Piece> chessBoard, Position position) {
-        if (chessBoard.containsKey(position)) {
-            return chessBoard.get(position).getName();
-        }
-        return DOT;
-    }
-
-    public boolean isGameEnded() {
-        return findChessBoard().isGameEnded();
-    }
-
-    public void deleteChessBoard() {
-        pieceDao.deleteAllPieces();
-        gameTurnDao.delete();
     }
 }
